LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/css-module/load-error-events-expected.txt

CONSOLE MESSAGE: TypeError: 'text/css' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: TypeError: 'text/css' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: TypeError: 'text/css' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: TypeError: 'text/css' is not a valid JavaScript MIME type.

FAIL inline, 200, parser-inserted assert_unreached: Error event should not be fired. Reached unreachable code
PASS inline, 404, parser-inserted
FAIL src, 200, parser-inserted assert_unreached: Error event should not be fired. Reached unreachable code
PASS src, 404, parser-inserted
FAIL src, 200, not parser-inserted assert_unreached: Error event should not be fired. Reached unreachable code
PASS src, 404, not parser-inserted
FAIL inline, 200, not parser-inserted assert_unreached: Error event should not be fired. Reached unreachable code
PASS inline, 404, not parser-inserted
